refactor: decompose LoggerUI Program.cs (752→38+74+183+317+73) into focused classes
Extract LogModels, LogEntryParser, LogEntryFormatter, PipeLogReader from monolithic Program.cs.
